requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pypdf2==3.0.1
pytesseract==0.3.10
pillow==10.1.0
numpy==1.24.3
# faiss-cpu is optional on Windows — only install on non-Windows platforms
faiss-cpu==1.7.4; sys_platform != "win32"
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
